step5/requirements.txt
azure-functions
azure-durable-functions
azure-storage-blob